LibGfx: Move implementation code for Tables.h into Tables.cpp
